move useAutosuggestMenu to a hooks file
